pyproject.toml
[project]
name = "hugo-trainsh-tools"
version = "0.1.0"
description = "Utility tools for the hugo-trainsh theme (preview image generator)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.5",
  "pillow>=10.4.0",
]

[tool.typer]
pretty_exceptions_enable = true


